chore: rename voting upgrade scripts
Use common omnibuses naming pattern
